+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+0.1.1 (2022-11-08)
+------------------
+
+* Define the first default config file and its schema.
+* First full run with the default config file.
+* Improve command line client.
+
 0.1.0 (2022-11-03)
 ------------------